+++ lemur/plugins/lemur_acme/tests/test_acme_handler.py
@@ -3,21 +3,11 @@ from unittest.mock import patch, Mock
 
 from cryptography.x509 import DNSName
 from lemur.plugins.lemur_acme import plugin
-from mock import MagicMock
 
 
 class TestAcmeHandler(unittest.TestCase):
-    @patch("lemur.plugins.lemur_acme.plugin.dns_provider_service")
-    def setUp(self, mock_dns_provider_service):
+    def setUp(self):
         self.acme = plugin.AcmeHandler()
-        mock_dns_provider = Mock()
-        mock_dns_provider.name = "cloudflare"
-        mock_dns_provider.credentials = "{}"
-        mock_dns_provider.provider_type = "cloudflare"
-        self.acme.dns_providers_for_domain = {
-            "www.test.com": [mock_dns_provider],
-            "test.fakedomain.net": [mock_dns_provider],
-        }
 
     def test_strip_wildcard(self):
         expected = ("example.com", False)
@@ -84,111 +74,4 @@ class TestAcmeHandler(unittest.TestCase):
         result = self.acme.get_domains(options)
         self.assertEqual(
             result, [options["common_name"], "test2.netflix.net"]
-        )
-
-    @patch(
-        "lemur.plugins.lemur_acme.plugin.AcmeHandler.start_dns_challenge",
-        return_value="test",
-    )
-    def test_get_authorizations(self, mock_start_dns_challenge):
-        mock_order = Mock()
-        mock_order.body.identifiers = []
-        mock_domain = Mock()
-        mock_order.body.identifiers.append(mock_domain)
-        mock_order_info = Mock()
-        mock_order_info.account_number = 1
-        mock_order_info.domains = ["test.fakedomain.net"]
-        result = self.acme.get_authorizations(
-            "acme_client", mock_order, mock_order_info
-        )
-        self.assertEqual(result, ["test"])
-
-    @patch(
-        "lemur.plugins.lemur_acme.plugin.AcmeHandler.complete_dns_challenge",
-        return_value="test",
-    )
-    def test_finalize_authorizations(self, mock_complete_dns_challenge):
-        mock_authz = []
-        mock_authz_record = MagicMock()
-        mock_authz_record.authz = Mock()
-        mock_authz_record.change_id = 1
-        mock_authz_record.dns_challenge.validation_domain_name = Mock()
-        mock_authz_record.dns_challenge.validation = Mock()
-        mock_authz.append(mock_authz_record)
-        mock_dns_provider = Mock()
-        mock_dns_provider.delete_txt_record = Mock()
-
-        mock_acme_client = Mock()
-        result = self.acme.finalize_authorizations(mock_acme_client, mock_authz)
-        self.assertEqual(result, mock_authz)
-
-    @patch("lemur.plugins.lemur_acme.plugin.AcmeHandler.setup_acme_client")
-    @patch("lemur.plugins.lemur_acme.plugin.current_app")
-    @patch("lemur.plugins.lemur_acme.plugin.authorization_service")
-    @patch("lemur.plugins.lemur_acme.plugin.dns_provider_service")
-    @patch("lemur.plugins.lemur_acme.plugin.AcmeHandler.get_authorizations")
-    @patch("lemur.plugins.lemur_acme.plugin.AcmeHandler.finalize_authorizations")
-    @patch("lemur.plugins.lemur_acme.plugin.AcmeHandler.request_certificate")
-    def test_get_ordered_certificate(
-            self,
-            mock_request_certificate,
-            mock_finalize_authorizations,
-            mock_get_authorizations,
-            mock_dns_provider_service,
-            mock_authorization_service,
-            mock_current_app,
-            mock_acme,
-    ):
-        mock_client = Mock()
-        mock_acme.return_value = (mock_client, "")
-        mock_request_certificate.return_value = ("pem_certificate", "chain")
-
-        mock_cert = Mock()
-        mock_cert.external_id = 1
-
-        provider = plugin.ACMEIssuerPlugin()
-        provider.get_dns_provider = Mock()
-        result = provider.get_ordered_certificate(mock_cert)
-        self.assertEqual(
-            result, {"body": "pem_certificate", "chain": "chain", "external_id": "1"}
-        )
-
-    @patch("lemur.plugins.lemur_acme.plugin.AcmeHandler.setup_acme_client")
-    @patch("lemur.plugins.lemur_acme.plugin.current_app")
-    @patch("lemur.plugins.lemur_acme.plugin.authorization_service")
-    @patch("lemur.plugins.lemur_acme.plugin.dns_provider_service")
-    @patch("lemur.plugins.lemur_acme.plugin.AcmeHandler.get_authorizations")
-    @patch("lemur.plugins.lemur_acme.plugin.AcmeHandler.finalize_authorizations")
-    @patch("lemur.plugins.lemur_acme.plugin.AcmeHandler.request_certificate")
-    def test_get_ordered_certificates(
-            self,
-            mock_request_certificate,
-            mock_finalize_authorizations,
-            mock_get_authorizations,
-            mock_dns_provider_service,
-            mock_authorization_service,
-            mock_current_app,
-            mock_acme,
-    ):
-        mock_client = Mock()
-        mock_acme.return_value = (mock_client, "")
-        mock_request_certificate.return_value = ("pem_certificate", "chain")
-
-        mock_cert = Mock()
-        mock_cert.external_id = 1
-
-        mock_cert2 = Mock()
-        mock_cert2.external_id = 2
-
-        provider = plugin.ACMEIssuerPlugin()
-        provider.get_dns_provider = Mock()
-        result = provider.get_ordered_certificates([mock_cert, mock_cert2])
-        self.assertEqual(len(result), 2)
-        self.assertEqual(
-            result[0]["cert"],
-            {"body": "pem_certificate", "chain": "chain", "external_id": "1"},
-        )
-        self.assertEqual(
-            result[1]["cert"],
-            {"body": "pem_certificate", "chain": "chain", "external_id": "2"},
         )
